Move a number of functions to blob_parser.py and helpers.py to make __init__.py more readable.
